requirements.txt
# Основные зависимости
pygame>=2.4.0
pillow>=10.0.0

# Опциональные зависимости
# numpy>=1.24.0
# matplotlib>=3.7.0
